lib: sbi: add initial context management component
The context management component in OpenSBI provides basic CPU
context initialization and management routines based on existing
domains in OpenSBI. The context management component was
originally designed to support a kind of security domain that could
yield CPU cores and be scheduled on demand, in which the UEFI
Secure Variable service runs.

The patch also provides an addition to the OpenSBI domain to provide
claims for this special security domain. Context management follows
the requirements for inter-domain isolation.

Signed-off-by: Qingyu Shang <2931013282@sjtu.edu.cn>